requirements.txt
mysql-connector-python>=8.0.0
Faker>=15.0.0
